backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic[email]>=2.5.3
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.20
httpx>=0.26.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
